refactor(hub): split DictationHub by domain (#970)
Reduces DictationHub from 586 LOC with 9 ctor deps and 21 mixed-domain
public methods to a 125-LOC wire-protocol facade with 3 ctor deps.
Domain logic moves to three focused command services under
Hubs/Services/, each ≤ 5 ctor deps with a single concern:

- IRemoteRecordingCommands (6 deps): dictation toggles, clipboard,
  keyboard helpers, ClearText. 12 methods all concerned with what the
  user types / dictates.
- IRemoteDesktopCommands (4 deps + logger): workspace switch, window
  close/activate, gtk-launch fallback, SnapFocusedWindowRight. 5
  methods all concerned with windows/workspaces.
- IRemoteScreenshotCommands (logger only): fresh-screenshot check and
  the insert-screenshot-path subprocess. 2 methods.

Method names on the hub are preserved verbatim, so the Remote Control
JavaScript client (wwwroot/remote.js) and existing IHubContext
consumers (DictationWorker, DesktopMonitorBroadcastWorker) continue
working against the same wire contract — no client-side or event-
broadcast changes needed. Splitting into three separate SignalR hubs
would have required rewriting the JS client and re-routing the 6+
event broadcasts that currently fan out via IHubContext<DictationHub>;
this refactor satisfies SRP / ISP without that wire risk.

The three command services are unit-testable without SignalR plumbing,
which unblocks real coverage for the previously-untested hub logic.
